# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Diario La Capital 2020",
        "contact": {
            "name": "Programador Principal",
            "email": "joseascurra123@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "operationId": "App\\Http\\Controllers\\CategoriasController::index",
                "responses": {
                    "200": {
                        "description": "Retorna todas las categorias de noticias existentes",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "operationId": "App\\Http\\Controllers\\CategoriasController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "En este parametro introducimos la id de la categoria a mostrar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna la categoria filtrada por el id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "No se especifico el id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/encuestas": {
            "get": {
                "tags": [
                    "Encuestas"
                ],
                "operationId": "App\\Http\\Controllers\\EncuestaController::index",
                "responses": {
                    "200": {
                        "description": "Obtiene todas las encuestas en formato json",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/encuestas/{id}": {
            "get": {
                "tags": [
                    "Encuestas"
                ],
                "operationId": "App\\Http\\Controllers\\EncuestaController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "La id unica de la encuesta a visualizar",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtiene la encuesta fitrada por el id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Encuestas"
                ],
                "operationId": "App\\Http\\Controllers\\EncuestaController::nuevaRespuesta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "La id unica de la encuesta a registrar la nueva respuesta",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "opcion",
                        "in": "query",
                        "description": "La opcion de la encuesta que elige el usuario",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "comentario",
                        "in": "query",
                        "description": "El comentario que realiza de acuerdo a la opcion elegida por el usuario",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve la respuesta realizada en formato json",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/lector": {
            "post": {
                "tags": [
                    "Lectores"
                ],
                "operationId": "App\\Http\\Controllers\\LectoresController::store",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Indica el nombre del lector",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellido",
                        "in": "query",
                        "description": "Indica el apellido del lector",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Indica el correo del lector",
                        "schema": {
                            "type": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna el lector registrado en formato json",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Noticias"
                ],
                "operationId": "App\\Http\\Controllers\\PostController::index",
                "responses": {
                    "200": {
                        "description": "Se obtienen todas las noticias del dia en un formato json",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Noticias"
                ],
                "operationId": "App\\Http\\Controllers\\PostController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "La id unica de la noticia a visualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtiene en formato json la noticia",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/vistas": {
            "get": {
                "tags": [
                    "Visitas"
                ],
                "operationId": "App\\Http\\Controllers\\VistaController::index",
                "responses": {
                    "200": {
                        "description": "Obtiene todas las visitas",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}